+++ src/app/views/brisca/brisca.component.ts
@@ -1,5 +1,6 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { interval, Subject, takeUntil } from 'rxjs';
 import { BottomControlsComponent } from '../../components/bottom-controls/bottom-controls.component';
 import { RoundInfoComponent } from '../../components/round-info/round-info.component';
 import { LOCAL_STORE_KEYS } from '../../constants/local-storage-keys';
@@ -23,13 +24,16 @@ const BRISCA_FLAGS = ['brisca', 'game:localStorageSave'] as const; // as Flag[]
   templateUrl: './brisca.component.html',
   styleUrls: ['./brisca.component.scss'],
 })
-export class BriscaComponent implements OnInit {
+export class BriscaComponent implements OnInit, OnDestroy {
   private readonly CLICKS_TO_DELETE_POINT = 5;
   public readonly BULLET_INTERVAL = 5;
+
+  private finishSubscriptions$ = new Subject<void>();
 
   public playersInfo: PlayerInfo[] = [];
   public showDeleteBanner = true;
   public gameService: GameService & GameServiceWithFlags<(typeof BRISCA_FLAGS)[number]>;
+  public playedTime: string;
 
   public constructor(gameHolderService: GameHolderService) {
     if (!gameHolderService.service.isGameServiceWithFlags(BRISCA_FLAGS as unknown as Flag[])) {
@@ -56,6 +60,9 @@ export class BriscaComponent implements OnInit {
         deletePointProgress: 0,
       };
     });
+
+    this.setPlayedTime();
+    interval(1000).pipe(takeUntil(this.finishSubscriptions$)).subscribe(() => this.setPlayedTime());
   }
 
   private checkShowDeleteBanner() {
@@ -85,5 +92,21 @@ export class BriscaComponent implements OnInit {
 
   public arrayOf(n: number): number[] {
     return intervalArray(n);
+  }
+
+  private setPlayedTime() {
+    const timeGameStarted = +localStorage.getItem(LOCAL_STORE_KEYS.TIME_GAME_STARTS)!;
+    const elapsedSeconds = (Date.now() - timeGameStarted) / 1e3;
+    const seconds = Math.trunc(elapsedSeconds % 60);
+    const minutes = Math.trunc((elapsedSeconds / 60) % 60);
+    const hours = Math.trunc(elapsedSeconds / 3600);
+    
+    this.playedTime = `${hours}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`
+    return;
+  }
+
+  public ngOnDestroy(): void {
+    this.finishSubscriptions$.next();
+    this.finishSubscriptions$.complete();
   }
 }