+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased version
+
+Added playing time to brisca game.
+
 ## 2.1.2
 
 Fix bug after editing number of players in game config when game has already started. Keep scores of each player.